Module aws_sdk_marketplacemetering::error

Modules

customer_not_entitled_exception

See CustomerNotEntitledException

disabled_api_exception

See DisabledApiException

duplicate_request_exception

See DuplicateRequestException

expired_token_exception

See ExpiredTokenException

internal_service_error_exception

See InternalServiceErrorException

invalid_customer_identifier_exception

See InvalidCustomerIdentifierException

invalid_endpoint_region_exception

See InvalidEndpointRegionException

invalid_product_code_exception

See InvalidProductCodeException

invalid_public_key_version_exception

See InvalidPublicKeyVersionException

invalid_region_exception

See InvalidRegionException

invalid_tag_exception

See InvalidTagException

invalid_token_exception

See InvalidTokenException

invalid_usage_allocations_exception

See InvalidUsageAllocationsException

invalid_usage_dimension_exception

See InvalidUsageDimensionException

platform_not_supported_exception

See PlatformNotSupportedException

throttling_exception

See ThrottlingException

timestamp_out_of_bounds_exception

See TimestampOutOfBoundsException

Structs

BatchMeterUsageError

CustomerNotEntitledException

Exception thrown when the customer does not have a valid subscription for the

DisabledApiException

The API is disabled in the Region.

DuplicateRequestException

A metering record has already been emitted by the same EC2 instance, ECS task, or

ExpiredTokenException

The submitted registration token has expired. This can happen if the buyer's

InternalServiceErrorException

An internal error has occurred. Retry your request. If the problem persists, post a

InvalidCustomerIdentifierException

You have metered usage for a CustomerIdentifier that does not exist.

InvalidEndpointRegionException

The endpoint being called is in a AWS Region different from your EC2 instance, ECS

InvalidProductCodeException

The product code passed does not match the product code used for publishing the

InvalidPublicKeyVersionException

Public Key version is invalid.

InvalidRegionException

RegisterUsage must be called in the same AWS Region the ECS task was launched in.

InvalidTagException

The tag is invalid, or the number of tags is greater than 5.

InvalidTokenException

Registration token is invalid.

InvalidUsageAllocationsException

The usage allocation objects are invalid, or the number of allocations is greater

InvalidUsageDimensionException

The usage dimension does not match one of the UsageDimensions associated with

MeterUsageError

PlatformNotSupportedException

AWS Marketplace does not support metering usage from the underlying platform.

RegisterUsageError

ResolveCustomerError

ThrottlingException

The calls to the API are throttled.

TimestampOutOfBoundsException

The timestamp value passed in the meterUsage() is out of allowed range.

Enums

BatchMeterUsageErrorKind

MeterUsageErrorKind

RegisterUsageErrorKind

ResolveCustomerErrorKind